demo/requirements.txt
streamlit
opencv-python
numpy
yolov5
Pillow

